+++ packages/spotlight/CHANGELOG.md
@@ -11,6 +11,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 ### Fixed
 
 - `spotlight` to handle non-5-way keys correctly to focus on next 5-way keys
+- `spotlight/Spottable` to forward `onMouseEnter` and `onMouseLeave`
 
 ## [1.12.1] - 2017-11-07