+++ packages/spotlight/CHANGELOG.md
@@ -10,12 +10,13 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
+- `spotlight/Spottable` to forward `onMouseEnter` and `onMouseLeave`
+
 ## [1.12.2] - 2017-11-15
 
 ### Fixed
 
 - `spotlight` to handle non-5-way keys correctly to focus on next 5-way keys
-- `spotlight/Spottable` to forward `onMouseEnter` and `onMouseLeave`
 
 ## [1.12.1] - 2017-11-07