+++ packages/spotlight/CHANGELOG.md
@@ -10,6 +10,10 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
+## [1.12.2] - 2017-11-15
+
+### Fixed
+
 - `spotlight` to handle non-5-way keys correctly to focus on next 5-way keys
 - `spotlight/Spottable` to forward `onMouseEnter` and `onMouseLeave`